feat(psychoakustik): gemeinsames Front-End + 6-Muster-Synergie (Qualität × Performance)
- NEU dsp/psychoacoustic_frame.py: EIN Frame (STFT 2048/512 + Fluss-Hüllkurve
  hop128 + Rectangular-Bark + Maskierungsschwelle ISO 11172-3) statt ~25
  getrennter STFT-Welten; Audibility-Routinen (Hörordnung §4) für Phasen
- einladungs_gate: vektorisierte Ganzsignal-Fenster statt 3 Fenster-Schleifen
  (~3x schneller; Roughness kalibriert: Sinus 0.04 / 70-Hz-AM 0.99)
- phase_02/phase_05: Audibility-Early-Termination (Hum/Rumble unter
  Maskierungsschwelle -> kein Eingriff, Budget 0 / HPF-Skip)
- unified_restorer_v3: Wohlklang-Garantie versucht Output-Blend vor Voll-Re-Run
  (gleicher MUSHRA-Proxy; Re-Run nur bei unzureichendem Blend)
- phase_04: Shelving-Pfad float32-native (Muster 5)
- +7 Frame-Unit-Tests; 88 Gate/Phasen-Tests, 331 UV3-Tests, Mypy-Gate 0, alle Datei-Gates grün

+++ backend/core/phases/phase_04_eq_correction.py
@@ -950,8 +950,11 @@ class EQCorrectionPhase(PhaseInterface):
         if not eq_corrections:
             return np.nan_to_num(audio, nan=0.0, posinf=0.0, neginf=0.0)  # type: ignore[no-any-return]  # No significant transport shift
 
-        # Apply corrections as smooth Butterworth shelving/peaking filters
-        result = np.asarray(audio, dtype=np.float64)
+        # Apply corrections as smooth Butterworth shelving/peaking filters.
+        # §Muster 5: float32-native — Shelving-Filter (1. Ordnung) sind in float32
+        # numerisch stabil; float64-Rundungsreisen verdoppelten nur den
+        # Speicherdurchsatz und erzwangen eine zweite Quantisierung.
+        result = np.asarray(audio, dtype=np.float32)
         try:
             from scipy.signal import butter  # pylint: disable=import-outside-toplevel
